feat: introduce project coloring
This commit introduce a way to differentiate project by color. This will
be used by cli for coloring (but is not used in this commit).

Creating a task with a unique project name will auto-assign the project
a color, following tasks sharing this project name won’t have to repeat
this action.

• project colors are randomized but I added some incentive to use colors
that are further away.
• no way to modify a project color in this commit

+++ libs/cli/commands/playground.rs
@@ -294,7 +294,7 @@ pub async fn handle(command: Command, _proxy: O324ServiceProxy<'_>) -> command_e
         let event_key = (event.start.timestamp(), event.end.timestamp());
         if let std::collections::hash_map::Entry::Vacant(e) = event_hues.entry(event_key) {
             // TODO
-            e.insert(200 as f32);
+            e.insert(200_f32);
             //if let Some(hue) = hue_generator.next_number() {
             //    e.insert(hue as f32);
             //}
